Fix infinite recursion bug causing stack overflow
If a metatable was being copied that had references to the same
metatable, the DeepCopy() method would recurse infinitely.

The simple solution is the let copies share a metatable with the
original. This is acceptable because the use case for this method is to
copy classes. There is no need to have a clone of the metatable because
the metatable does not need to be modified.