Add ExtraObjects to TurnCounter
This enables the scenario designer to place relevant game objects next
to the Turn Counter. For example, they could place an artillery or
airstrike card next to its corresponding event.

Note: This feature was rushed to production because I need to release
the next expansion tonight.

Tested on the upcoming Sainte Mere Eglise content